requirements.txt
requests
requests-toolbelt
cachecontrol
six
future
matplotlib>=3.6.2
numpy>=1.24.1
scikit-learn
scikit-image
pandas
opencv_contrib_python>=4.7.0.68
openslide_python>=1.2.0
Pillow>=9.5.0
pyvips>=2.2.1
scipy>=1.10.1
SimpleITK>=2.2.1
torch>=1.13.1
torchvision>=0.14.1
torchsummary
torchio
